Created a class to hold rx-operation objects
