+++ rxbindings/src/main/java/com/amplifyframework/rx/RxApiBinding.java
@@ -16,7 +16,6 @@
 package com.amplifyframework.rx;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 
 import com.amplifyframework.api.ApiCategory;
@@ -27,21 +26,17 @@ import com.amplifyframework.api.graphql.GraphQLResponse;
 import com.amplifyframework.api.rest.RestOptions;
 import com.amplifyframework.api.rest.RestResponse;
 import com.amplifyframework.core.Amplify;
-import com.amplifyframework.core.Consumer;
-import com.amplifyframework.core.async.Cancelable;
 import com.amplifyframework.rx.RxAdapters.CancelableBehaviors;
-
-import java.util.Objects;
+import com.amplifyframework.rx.RxOperations.RxSubscriptionOperation;
 
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.core.Single;
-import io.reactivex.rxjava3.subjects.BehaviorSubject;
 
 /**
  * An implementation of the RxApiCategoryBehavior which satisfies the API contract by wrapping
  * {@link ApiCategoryBehavior} in Rx primitives.
  */
-final class RxApiBinding implements RxApiCategoryBehavior {
+public final class RxApiBinding implements RxApiCategoryBehavior {
     private final ApiCategoryBehavior api;
 
     RxApiBinding() {
@@ -177,132 +172,4 @@ final class RxApiBinding implements RxApiCategoryBehavior {
         return CancelableBehaviors.toObservable(method);
     }
 
-    /**
-     * A class that represents a subscription operation and exposes
-     * observables for consumers to listen to subscription data and
-     * status events.
-     * @param <T> The type representing the subscription data.
-     */
-    public static final class RxSubscriptionOperation<T> implements Cancelable {
-        private BehaviorSubject<ConnectionStateEvent> connectionStateSubject;
-        private Observable<T> subscriptionData;
-        private Cancelable amplifyOperation;
-        private OnConnectedConsumer onConnected;
-
-        RxSubscriptionOperation(CancelableBehaviors.StreamEmitter<String, T, ApiException> callbacks) {
-            onConnected = new OnConnectedConsumer();
-            connectionStateSubject = BehaviorSubject.create();
-            subscriptionData = Observable.create(emitter -> {
-                amplifyOperation = callbacks.streamTo(onConnected::accept,
-                                                      emitter::onNext,
-                                                      emitter::onError,
-                                                      emitter::onComplete);
-            });
-        }
-
-        /**
-         * Returns an {@link Observable} which consumers can use to
-         * retrieve data received by the subscription operation.
-         * @return Reference to the {@link Observable} with subscription data.
-         */
-        public Observable<T> observeSubscriptionData() {
-            return subscriptionData;
-        }
-
-        /**
-         * Returns an {@link Observable} which consumers can use to
-         * receive notfication about the status of the subscription connection. Currently,
-         * only {@link ConnectionState#CONNECTED} is emitted.
-         * @return Reference to the {@link Observable} that receives connection events.
-         */
-        public Observable<ConnectionStateEvent> observeConnectionState() {
-            return connectionStateSubject;
-        }
-
-        @Override
-        public void cancel() {
-            connectionStateSubject.onComplete();
-            if (amplifyOperation != null) {
-                amplifyOperation.cancel();
-            }
-        }
-
-        /**
-         * Enum representing connection states of a
-         * subscription operation.
-         */
-        public enum ConnectionState {
-            /**
-             * The subscription successfully established a connection and is
-             * ready to receive data.
-             */
-            CONNECTED
-        }
-
-        /**
-         * Describes events emitted by the {@link RxSubscriptionOperation} class.
-         */
-        static class ConnectionStateEvent {
-            private ConnectionState connectionState;
-            private String subscriptionId;
-
-            ConnectionStateEvent(@NonNull ConnectionState connectionState,
-                                 @Nullable String subscriptionId) {
-                this.connectionState = connectionState;
-                this.subscriptionId = subscriptionId;
-            }
-
-            /**
-             * The connection state reported in the event.
-             * @return The {@link ConnectionState} representing the state of the connection.
-             */
-            @NonNull
-            public ConnectionState getConnectionState() {
-                return connectionState;
-            }
-
-            /**
-             * The subscriptionId associated with the event.
-             * @return The value of the subscriptionId.
-             */
-            @Nullable
-            public String getSubscriptionId() {
-                return subscriptionId;
-            }
-
-            @Override
-            public int hashCode() {
-                int result = connectionState.hashCode();
-                result = 31 * result + (subscriptionId != null ? subscriptionId.hashCode() : 0);
-                return result;
-            }
-
-            @Override
-            public boolean equals(@Nullable Object obj) {
-                if (this == obj) {
-                    return true;
-                } else if (obj == null || getClass() != obj.getClass()) {
-                    return false;
-                } else {
-                    ConnectionStateEvent privateNote = (ConnectionStateEvent) obj;
-                    return Objects.equals(getConnectionState(), privateNote.getConnectionState()) &&
-                        Objects.equals(getSubscriptionId(), privateNote.getSubscriptionId());
-                }
-            }
-
-            @NonNull
-            @Override
-            public String toString() {
-                return "ConnectionStateEvent{connectionState=" + connectionState + "," +
-                    "subscriptionId=" + subscriptionId + "}";
-            }
-        }
-
-        private final class OnConnectedConsumer implements Consumer<String> {
-            @Override
-            public void accept(@NonNull String subscriptionId) {
-                connectionStateSubject.onNext(new ConnectionStateEvent(ConnectionState.CONNECTED, subscriptionId));
-            }
-        }
-    }
 }
