+++ rxbindings/src/main/java/com/amplifyframework/rx/RxGraphQlBehavior.java
@@ -22,6 +22,7 @@ import com.amplifyframework.api.ApiException;
 import com.amplifyframework.api.graphql.GraphQLBehavior;
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.api.graphql.GraphQLResponse;
+import com.amplifyframework.rx.RxOperations.RxSubscriptionOperation;
 
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.core.Single;
@@ -154,7 +155,7 @@ public interface RxGraphQlBehavior {
      *         the GraphQL network subscription will be closed.
      */
     @NonNull
-    <T> RxApiBinding.RxSubscriptionOperation<GraphQLResponse<T>> subscribe(
+    <T> RxSubscriptionOperation<GraphQLResponse<T>> subscribe(
             @NonNull GraphQLRequest<T> graphQlRequest
     );
 
@@ -180,7 +181,7 @@ public interface RxGraphQlBehavior {
      *         the GraphQL network subscription will be closed.
      */
     @NonNull
-    <R> RxApiBinding.RxSubscriptionOperation<GraphQLResponse<R>> subscribe(
+    <R> RxSubscriptionOperation<GraphQLResponse<R>> subscribe(
             @NonNull String apiName,
             @NonNull GraphQLRequest<R> graphQlRequest
     );
